requirements.txt
numpy
sklearn
pandas
dill
torch
tqdm
pyyaml==5.1
python-telegram-bot
